UspeH

+++ server/src/main/java/com/server/studio27/controllers/StudentController.java
@@ -35,7 +35,7 @@ public class StudentController {
 
             for (Map<String, Object> row : rows) {
                 Map<String, Object> studentMap = new HashMap<>();
-                studentMap.put("studentId", ((Integer) row.get("studentId")).intValue());
+                studentMap.put("studentId", ((Long) row.get("studentId")).intValue());
                 studentMap.put("email", (String) row.get("email"));
                 studentMap.put("password", (String) row.get("password"));
                 studentMap.put("ime", (String) row.get("ime"));
@@ -51,11 +51,11 @@ public class StudentController {
                                     where studentId = ?
                                     """;
                 System.err.println("Sql2: " + SQL2);
-                List<Map<String, Object>> kursRows = jdbcTemplate.queryForList(SQL2, ((Integer) row.get("studentId")).intValue());
+                List<Map<String, Object>> kursRows = jdbcTemplate.queryForList(SQL2, ((Long) row.get("studentId")).intValue());
                 List<Map<String, Object>> kursevi = new ArrayList<>();
                 for (Map<String, Object> kursRow : kursRows) {
                     Map<String, Object> kursMap = new HashMap<>();
-                    kursMap.put("kursId", ((Integer) kursRow.get("kursId")).intValue());
+                    kursMap.put("kursId", ((Long) kursRow.get("kursId")).intValue());
                     kursMap.put("naziv", (String) kursRow.get("naziv"));
                     kursevi.add(kursMap);
                 }
